generated markdown README based on previous commit<div id="table-of-contents"> <h2>Table of Contents</h2> <div id="text-table-of-contents"> <ul> <li><a href="#sec-1">1. C++ Project Template</a> <ul> <li><a href="#sec-1-1">1.1. Introduction</a></li> <li><a href="#sec-1-2">1.2. How to use</a></li> <li><a href="#sec-1-3">1.3. How to contribute</a></li> <li><a href="#sec-1-4">1.4. Licence</a></li> </ul> </li> </ul> </div> </div>
A C++ Project Template.
Because you should concentrate on the important parts.

This is a project template for C++.
Making files, folders, CMakeLists.txt and all the boring stuff all done and ready to go.

1.  forking
    fork this project and use it. If you are really lazy.

2.  cloning
    if you do not want this project to appear as forked,
    clone this project and use it as you wish.

This project is open for contribution.
Send me push requests any time you have a good idea.
Please describe reasons for additions or fixes.

I would really appreciate contributions from experienced CMakeLists.txt users.

MIT License

Copyright (c) 2017 Red-Portal

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:<div id="table-of-contents">
<h2>Table of Contents</h2>
<div id="text-table-of-contents">
<ul>
<li><a href="#sec-1">1. C++ Project Template</a>
<ul>
<li><a href="#sec-1-1">1.1. Introduction</a></li>
<li><a href="#sec-1-2">1.2. How to use</a></li>
<li><a href="#sec-1-3">1.3. How to contribute</a></li>
<li><a href="#sec-1-4">1.4. Licence</a></li>
</ul>
</li>
</ul>
</div>
</div>

A C++ Project Template.
Because you should concentrate on the important parts.

This is a project template for C++.
Making files, folders, CMakeLists.txt and all the boring stuff all done and ready to go.

1.  forking
    fork this project and use it. If you are really lazy.

2.  cloning
    if you do not want this project to appear as forked,
    clone this project and use it as you wish.

This project is open for contribution.
Send me push requests any time you have a good idea.
Please describe reasons for additions or fixes.

I would really appreciate contributions from experienced CMakeLists.txt users.

MIT License

Copyright (c) 2017 Red-Portal

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions: